Creating a Compiler Pass
Most of the time, you won't need to create a compiler pass - you just need to understand how they work. But, we're diving deep, so let's make one! In AppBundle create a new DependencyInjection directory and inside of there a Compiler directory. I don't have to put it here, but this follows the core standard.

In here, create a new class called EarlyLoggingMessagePass. Remember how we logged a message as soon as the logger was created? We're going to do that again.

Compiler classes are pretty easy - just implement CompilerPassInterface and add the one method: process()

Now we should feel really comfortable: that's a ContainerBuilder object and we know all about him. It also has every service already defined inside. So we can say: {dollar_sign}definition = {dollar_sign}container->findDefinition('logger'). Now just add {dollar_sign}definition->addMethodCall() and pass it debug for the method, and an array with a single argument: Logger CREATED

And that's a functional compiler pass.

You can register this by overriding the build() method in AppBundle and adding it there. But that's too easy.

Instead, go to AppKernel and override buildContainer(). Call the parent method, then add {dollar_sign}container->addCompilerPass() and pass it a new EarlyLoggingMessagePass. And don't forget to return the {dollar_sign}container.

Ok, let's try it! Refresh! Click into the profiler then go to the logs tab. Under debug, there's the message! First on the list.

Phew! So you're now a master. The Container is all about Definition objects, which are populated from Yaml and XML files and then updated later in the dependency injection extension classes. If you're following this, go dive into the FrameworkBundle and see where the real core services come from And congrats, because now, you're a dependency-injection-asaurus!